# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaborative Whiteboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Room {
  id: string;
  name: string;
  createdAt: string;
}

export interface RoomInput {
  name?: string;
}

export type CanvasElementType = typeof CanvasElementType[keyof typeof CanvasElementType];


export const CanvasElementType = {
  pen: 'pen',
  rect: 'rect',
  ellipse: 'ellipse',
  arrow: 'arrow',
  text: 'text',
} as const;

export interface CanvasElement {
  id: string;
  roomId: string;
  type: CanvasElementType;
  data: string;
  createdAt: string;
}

export type RoomStatsByType = {
  pen?: number;
  rect?: number;
  ellipse?: number;
  arrow?: number;
  text?: number;
};

export interface RoomStats {
  totalElements: number;
  activeUsers: number;
  byType: RoomStatsByType;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}