# lib/api-zod/src/generated/types/canvasElement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaborative Whiteboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CanvasElementType } from './canvasElementType';

export interface CanvasElement {
  id: string;
  roomId: string;
  type: CanvasElementType;
  data: string;
  createdAt: string;
}