# lib/api-zod/src/generated/types/roomStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaborative Whiteboard API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomStatsByType } from './roomStatsByType';

export interface RoomStats {
  totalElements: number;
  activeUsers: number;
  byType: RoomStatsByType;
}